fix(cp): v2.0.1 hotfix — timeline filters, audit humaniser, widget ACL, float→int warning
Four post-v2.0.0 fixes surfaced from a production install.

* Timeline filter pills (System / Audit / Errors / Warnings / Info)
  now auto-submit the filter form on change with a short debounce,
  and toggle `.lb-pill--active` client-side for instant feedback.
  Hidden inputs moved from `style="display:none"` to `.lb-sr-only`
  so label-click semantics stay portable. Form tagged
  `data-lb-filter-form` so the shared empty-param stripper applies.

* Timeline audit rows now render via `AuditActionPresenter::label()`
  + `variant()` (matching the Audit Logs page) — "User updated"
  instead of `statamic.user.updated`, correct chip colour, and a
  truncated `changeSummary()` hint inline on update events. Raw
  event string preserved as chip tooltip so ?action= filters on
  the Audit page keep working.

* `LogbookDashboardData::lastErrorAt()` was feeding the return
  value of `Carbon::diffInMinutes()` (float in Carbon 3) into the
  `int`-typed `humaniseMinutes()` helper, triggering
  "Implicit conversion from float to int loses precision" on
  PHP 8.3+ — which got recorded as a warning row in our own
  logbook every dashboard render. Explicit `(int)` cast at the
  call site.

* Dashboard widgets now honour the `view logbook` permission.
  `LogbookStatsWidget`, `LogbookTrendsWidget`, and
  `LogbookPulseWidget` all short-circuit `html()` to a marker
  `<div class="lb-widget-gated" hidden>` when the current CP
  user can't view the logbook. The shipped stylesheet uses
  `:has(.lb-widget-gated)` to collapse the enclosing dashboard
  card entirely — no title, no surround, no data leak.

Bumps package to 2.0.1 and records all four fixes in CHANGELOG.

+++ src/Support/LogbookDashboardData.php
@@ -515,7 +515,12 @@ class LogbookDashboardData
         }
 
         $at = Carbon::parse($row);
-        $diffMinutes = max(0, $at->diffInMinutes(now()));
+        // Carbon 3 (Laravel 11+) returns a float here; Carbon 2 returns an int.
+        // humaniseMinutes() is typed `int`, so without the cast PHP 8.3+ emits
+        // an "Implicit conversion from float to int loses precision" warning
+        // every time the widget renders — which, ironically, gets recorded as
+        // a warning row in our own logbook.
+        $diffMinutes = (int) max(0, $at->diffInMinutes(now()));
 
         return [
             'at' => $at,
